update logical router in the background task
This branch adds an implementation to the `todo!`'d background task for
updating the logical router's shared state. The background task updates
the set of backends, distribution, and routes when the profile watch
changes. Code for actually generating different parts of the stack is
split out into functions so that it can be shared between the
`new_service` impl that generates the initial state of the router, and
the background task itself.

It might be potentially worth investigating a way to avoid rebuilding
parts of the profile when only some of the profile changes (e.g. if the
set of targets remains constant but the set of routes changes, only
rebuild the routes, and so on). However, I punted on this for now as it
could potentially introduce a lot of complexity.