Propagate backpressure from buffers when in failfast (#2091)
When a buffer is in failfast, all requests are eagerly drained from the buffer
so that callers have no ability to detect that the buffer is unavailable.

This change adds a new type of failfast layer, `FailFast::layer_gated`, that
can be used to wrap buffers such that the buffer appears unavailable as soon as
failfast is activated.

This change also adds a new `LoadShed` layer that can be used to fail requests
when a buffer is at capacity. Moving this load shedding 'up' the stack will
allow the `Distribute` middleware to consider alternate backend services when
one service is in failfast.