stack: add `SpawnWatch` middleware
This branch adds an implementation of a middleware that watches a
`watch::Receiver` of targets and builds a new service when the watch is
updated. The current state of the inner service is broadcast over a
second watch to any clones of the `SpawnWatch` service.

This is the same code added in PR #2100, but rebased onto `main`, with
documentation and tests added.