change logical router to return an error when no route is found
Currently, the logical router in #2095 hits a TODO and panics when no
route matches a request. The Gateway API specifies that if no route in
an HTTPRoute matches a request, a 404 error must be returned.
ServiceProfiles, on the other hand, simply don't implement any per-route
behavior when a request doesn't match any routes.

PR #2108 added default routes when policies are generated from
ServiceProfiles. Therefore, we can have the router fail when no route is
matched, because ServiceProfiles will always contain a matching route.
This branch changes the router behavior so that an error is returned if
a request doesn't match any of its routes, and handles that error by
returning a 404 response.

Depends on #2108.