fix use of default profile in tests
Currently, the `Profile` type has a `Default` implementation which
returns an empty `Profile`. This is only used in tests.

In commit e122dd015a49a0b6ff6f1d1f0d98325e5e6d002f, responsibility for
handling profiles with empty lists of targets or empty routes was moved
from the logical router to the code for converting protobuf profiles
into the proxy's internal `Profile` type. This means that an instance of
the `Profile` type which has neither a direct endpoint override or a
list of backends is now invalid and causes the router to fail.
Therefore, tests which construct `Profile`s using `Profile::default()`
now fail, because the default profile is invalid.

This branch fixes these tests by doing the following:

- Remove the `Default` implementation for `Profile`, and replace it with
  a `profile::empty()` function in the `linkerd-test` crate. Having a
  way to generate an empty profile easily is still useful, because we
  may want to override some fields of a profile and leave the rest
  empty, but renaming the function to `empty()` makes it clearer what
  we're doing. Since the `Default` impl was only used in tests, it makes
  more sense for the code to live in the test-support crate.

  Perhaps it would be better to name the function `invalid()`, so that
  when reading tests, it's obvious that the profile will not be valid if
  it's constructed using only that funtion?

- Change the `profile::with_addr()` function to correctly populate the
  profile's `targets` and `target_addrs` fields to match the way we
  would generate a `Profile` when constructing it from protobuf.

- Change tests that used `Profile::default()` without populating the
  endpoint override field to use `profile::with_addr()` instead, so that
  the profile is valid.