distribute: Add a backend cache
`NewDistribute` will be used to implement per-route traffic distribution
policies so that, for instance, each request route may use a different
traffic split. In this setup, we want each instance--re-constructed as
configuration changes--to reuse a common set of concrete services so
that load balancer state need not be lost when a policy changes.

This change adds a `distribute::BackendCache` module that produces a
`NewDistribute` for an updated, cached set of backends.