stack: Eliminate the `UpdateWatch` trait
We've learned that there's no need for a specialized `UpdateWatch`
trait. Instead, this change uses a hierarchy of `NewService` instances.
The first `NewService` builds an inner stack for each watch, so that we
can isolate per-watch scope; and the second builds an inner stack each
time the watch updates.

The `NewSpawnWatch::layer_into` builder alters the update-level target
type by combining the watch-level and update-level target types.

This is intended to combine with #2122.