fix empty distributions for routes without backends
Currently, when constructing the `RouteParams` for a `Profile` which
does not have a list of targets, we add the logical address as a
backend, but we still generate an empty `Distribution`. Empty
`Distribution`s never become ready, so these profiles will hang forever.

This commit fixes this by also populating the `Distribution` with the
logical address with a weight of 1 when the profile has no backends.