Parameterize the load balancer stack (#2142)
The load balancer/discovery modules in the concrete stacks use
hardcoded ewma parameters.

In preparation for making these parameters discoverable via the
API, this change modifies the balancer to accept this configuration
as a target param.

Furthermore, this change cleans up the balancer API to simplify the
way that we convert Resolve resolutions to Discover streams for
the balancer: the discovery buffer is spawned immediately instead
of using a MakeService. In practice, we flatten the 'make' future
into the discovery stream in the readycache, anyway; so there's no
functional difference.

This change removes the discovery watchdog timeout. We've never
seen it triggered in practice.