stack: add `AnnotateError` middleware
This branch introduces a new `AnnotateError` middleware in
`linkerd-stack`, for adding context to errors returned by a wrapped
`Service`.

The `annotate_error` module contains a new trait, `ErrorContext`, which
is implemented by types that provide context to `Error`s by wrapping
them with a new `Error` type, and the `AnnotateError` middleware itself.
The middleware consists of a `NewService` which extracts a `Param`
implementing `ErrorContext` and produces a `Service`, which wraps any
errors returned by the inner `Service` using that `ErrorContext`.

Additionally, `NewAnnotateError` may be constructed with an
`ExtractParam` implementation that configures how an `ErrorContext`
implementation is produced from a target type, when the `Param`s
themselves do not implement `ErrorContext`. Finally, a pre-made
implementation of `ErrorContext`, called `FromTarget`, that produces new
errors when the error type implements `From<(&T, Error)>` is provided.
In this case, the error type can implement `From` with the appropriate
target type, and the `annotate_error::layer_from_target` constructor can
be used, without requiring an `ExtractParam` implementation to be
written for each stack.