metrics: add response classification to latency metrics
Currently, the documentation for proxy metrics [states][1] that all
`response_`-prefixed metrics are labeled with response classifications.
However, the `response_latency_ms` histograms lack this label --- they
are labeled with status codes, but not with response classifications or
gRPC status labels.

This is because response latencies are calculated when the last `DATA`
frame of a response body is received, rather than when a `TRAILERS`
frame is received or the body stream ends (depending on whether the
response includes trailers). Response classifications, however, cannot
be determined until a `TRAILERS` frame is received or the body stream
terminates, as a `TRAILERS` frame may contain a `grpc-status` trailer
that influences response classification.

This branch adds response classification labels to the
`response_latency_ms` histogram metrics. In order to do so, it was
necessary to change how response latencies are recorded slightly. Now,
the `ResponseBody` type still _calculates_ the latency when the last
`DATA` frame of a response is received, but the latency is not actually
recorded in the histogram until a response classification is recorded.
This way, the response latency histogram can be part of the
`ClassMetrics` structure rather than the `StatusMetrics` structure.

Fixes linkerd/linkerd2#9737

[1]: https://linkerd.io/2.12/reference/proxy-metrics/#response-labels