collapse `StatusMetrics` and `ClassMetrics` into one hashmap
Now that response latency metrics are labeled with a classification as
well as a status code, it is no longer necessary to maintain a hashmap
of metrics by status code which contains hashmaps of metrics by
classification, since all response metrics are now labeled with both a
status and a classification. Therefore, we can combine these two nested
hashmaps into a single map keyed by `(Option<Status>, Classification)`.
This makes the code a bit simpler and may be slightly more efficient due
to fewer pointer dereferences to traverse nested maps.