orig_proto: don't set `connection: close` on errors (#2171)
When the inbound proxy's `ClientRescue` layer sees a client connection
error for a connection initiated for an HTTP/1 request, it will add a
`connection: close` header to the synthesized error response. The
`ClientRescue` layer is not currently aware of whether an HTTP/1 request
was recieved from an upstream proxy as part of an orig-proto upgrade or
not. If the request _is_ part of an orig-proto upgrade, the server side
of the inbound proxy is an HTTP/2 server. When Hyper's HTTP/2 server
sees a response with a `connection: close` header, it emits a warning,
because `connection` headers are illegal in HTTP/2.

The inbound proxy *does* strip any headers set by the server that are
illegal in HTTP/2 when performing an orig-proto downgrade. However, this
stripping occurs only when handling a response received from the server
peer, and *not* on synthesized error responses generated by the proxy.
Therefore, if a connection-level error occurs on a downgraded request,
the inbound proxy will log an illegal header warning. This can be
confusing to users, as this is not the actual root cause of the error
(see linkerd/linkerd2#10090).

This branch changes the `ErrorRespond` middleware to only add
`connection: close` to synthesized HTTP/1 responses that are *not* part
of an orig-proto downgrade. This is accomplished by adding a request
extension which is set in the `orig_proto::Downgrade` layer indicating
that a request was downgraded to the original protocol. The 
`connection: close` header is now only added if the downgrade marker
extension is not present, and the warning is now no longer logged.

Fixes linkerd/linkerd2#10090