configure buffers from target `Param`s (#2173)
Buffer services in the proxy are constructed with a queue capacity and a
failfast timeout. Currently, these parameters are configured based on
the proxy's config structs (parsed from environment variables), and each
buffer `Layer` is created with a fixed buffer configuration depending on
which stack it is used in, so all buffers constructed by that stack will
have the same configuration. In the future, we would like to allow these
parameters to be configured per-target by the control plane.

In order to support configuring buffer parameters dynamically, this
branch introduces a `NewQueue` type, which implements `NewService` and
produces `Buffer` + `FailFast` services configured using a `QueueConfig`
`Param` extracted from the target. This will allow target types to
implement `Param<QueueConfig>` to dynamically configure the buffer
parameters.

In order to continue supporting the current state of the world (i.e.,
fixed buffer parameters) without adding them to every target type,
`NewQueue`'s `NewService` implementation is generic over an
`ExtractParam` implementation to control how the `QueueConfig` params
are extracted from targets. Convenience constructors are provided for
`()` as the `ExtractParam` type, expecting that `T: Param<QueueConfig>`,
and for `CloneParam` as the `ExtractParam` type, to clone the same set
of params every time a buffer is constructed. Currently, all `NewQueue`
layers constructed in the proxy use `CloneParam`, to continue
implementing the existing fixed configuration behavior.

Co-authored-by: Oliver Gould <ver@buoyant.io>