Update HTTP error responder to log version info
The HTTP error responder already knows whether a request is considered
HTTP/1, HTTP/2, or gRPC.

This change updates these log messages to include the relevant protocol
information for each failed request.

Furthermore, this change updates the `http::Version` enum to implement
`Debug` with full HTTP version strings like "HTTP/1" and "HTTP/2". We
then remove the version for debug spans, as it's needlessly verbose for
most uses. When debugging, we can identify the version based on specific
events in a span. It need not exist on every event on the span.