Add the `meshtls-boring-fips` feature flag (#2168)
The `boring` crate exposes a `fips` feature flag, but this feature
was not exposed by the `linkerd-meshtls-boring`. This change makes
it theoretically possible to build the proxy in this mode. This feature
is NOT tested in CI due to the fussiness of the required build
environment and longer-term maintenance concerns.

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>