add target metadata to error contexts (#2162)
In order to improve the proxy's debuggability, it's desirable that the
errors logged by the proxy and/or returned to HTTP clients include
relevant information about where in the proxy that error occurred.
Currently, buffer layers are constructed with a string "name" parameter
that describes which part of the stack is behind that buffer, so that
the fail-fast errors returned by those buffers can contain the stack
layer's name. While this provides some useful context, it isn't a
complete solution, as this context is only added to fail-fast errors,
and not other errors such as I/O errors, Additionally, these names only
indicate where in the stack the error occurred, and don't provide
metadata describing *which instance* of that stack layer produced the
error --- e.g., there is no context describing the client address,
logical address, concrete address, or endpoint (as appropriate) that
produced the error.

Instead, we should annotate *all* errors with the following information:

- The name of the stack layer where the error occurred
- Target metadata including the logical, concrete, and endpoint
  addresses (as appropriate)

This branch uses the `AnnotateError` middleware added in PR #2158 to add
layers to the proxy which annotate errors with metadata extracted from targets.
Furthermore, this branch removes the existing name parameter from buffer
layers, and replaces them with the `AnnotateError` middleware, which is
used to add target metadata *and* a string name describing the stack
layer to all errors. Additionally, the various `HttpRescue`
implementations have been changed slightly, so that when a specific
cause is found for an error, the entire error chain is still formatted
in the HTTP response `l5d-error` header, rather than _just_ the cause.
This ensures that all context added to an error is included in the
formatted message, rather than just the root cause (which lacks target
metadata).