Cache discovery results independently of proxy stacks  (#2195)
The current discovery module resolves profiles from the control plane
and injects the response into the inner stack. So, in order to cache
this result, the entire stack must be cached. This prevents us from
interacting with discovery in any stack that includes
higher-cardinality, peer-specific metadata. This limitation is
unnecessary and cumbersome to work around.

This change adds an `app::core::NewCachedDiscover` module that
internally manages resolutions independently of the proxy stack.

This change does **not** remove any other existing caches, though it's
expected that several of these caches are no longer necessary.

Co-authored-by: Eliza Weisman <eliza@buoyant.io>