gateway: Unify discovery for HTTP & opaque stacks (#2198)
In its normal sidecar configuration, the outbound stack does discovery
for each target ip:port. The result of this discovery will influences
protocol detection, etc.

In the gateway configuration, however, discovery is pushed into each
distinct stack so that lookups for HTTP services are deferred until
we've terminated the connection and wired it into the per-request
outbound stack.

This difference was primarily motivated by a limitation of the way that
discovery information is cached: because the outermost target includes
per-client metadata, we can't cache a discovery stack based on that
target type. #2195 removes this limitation so that we can update the
gateway stack to look more like a typical outbound stack (which will
allow for easier code reuse as new discovery types are introduced).

This change:

1. Splits the gateway stack into smaller functions for readability.
2. Extracts discovery from the HTTP and opaque stacks.
3. Renames gateway types for clarity.