http: Remove `Clone` requirement in servers (#2200)
The HTTP server currently shares an inner HTTP service across all
connections, requiring that the inner service is `Clone` (and therefore
buffered).

Furthermore, the HTTP server instruments requests with a `ClientHandle`
so that the inner service can signal connection teardown. This
client-specific context would, ideally, be available as the stack is
constructed and not only when processing requests.

To remove the `Clone` requirement, this change updates `ServerHttp` to
create a new inner service for each connection. It is then left up to
the inner stack to determine whether is provided via cloning or
otherwise.

This change does **not** change the shape of any existing stacks.
Instead, we add a `Stack::unlift_new` utility (the opposite of
`Stack::lift_new`). "Unlift" seems less ambiguous than "drop".