outbound: Make stacks generic over target types
We want to change the shape of the outbound stack--removing
`switch_logical` so that forwarding/balancing decisions are made in
concrete stacks. We also want to move away from using profiles
exclusively, adding new discovery types. This is complicated by our use
of fixed target types because it's difficult to make incremental
changes. Furthermore, a stack's requirements are not clearly documented.

This change addresses this as follows:

* The logical, concrete, and endpoint stacks are generic over their
  input types. Each stack enumerates required parameters from these
  targets.
* Each of these stacks declare module-local types that are passed to
  inner services. These types must expose params to satisfy all of the
  requirements of inner stacks.
* `tcp::concrete` and `tcp::logical` have been renamed to
  `opaque::concrete` and `opaque::logical` to reflect that they are an
  alternative to the http stack. The HTTP stack continues to incorporate
  the `tcp::endpoint` stack. The stack helpers have been renamed
  accordingly.