downgrade `miniz_oxide` from yanked 0.6.4 to 0.6.2
Versions 0.6.4 and 0.6.3 of the `miniz_oxide` crate have been yanked
from crates.io, due to an accidental MSRV violation (see
Frommi/miniz_oxide#135). `miniz_oxide` is a dependency of `flate2`,
which is used by `linkerd-app-integration` for testing gzip metrics
compression.

Because the lockfile currently contains a yanked version of
`miniz_oxide`, our `cargo deny` checks are failing on CI. This PR
downgrades the `miniz_oxide` dependency to 0.6.2, which is not yanked.
This should fix CI.