Use a cascading stack with protocol detection (#2221)
The protocol detection stack does not easily allow for its inner stack
to hold any per-target state because it commingles the (async) detection
result with the original target.

This change updates the detection module to use a cascading inner stack
so that the caller may determine how to use these targets. All
implementations lift the original targets to preserve the prior
behavior.