Update `linkerd2-proxy-api` to v0.8.0 (#2223)
v0.8.0 of `linkerd2-proxy-api` adds a `ProtocolHint::Opaque` variant to
the Destination API to indicate that an endpoint will not perform
application-level processing. This branch updates the proxy to depend on
v0.8.0. Currently, the `Opaque` hint does not change the proxy's
behavior; this will be added in a follow-up branch.