outbound: Refactor stack target types
To support a new configuration API, we want to update the outbound stack
to compose the following stacks:

* Routing discovery;
* Protocol switching;
* Protocol-specific logical routing;
* Concrete forwarding / load balancing;
* Endpoint dispatching.

To do this, we update the existing stacks so that:

* each stack accepts generic input targets (with enumerated `Param`
  requirements); and
* each module defines all necessary internal target types, including
  those passed to inner stacks.

This decouples stack configuration from the specifics of discovery:
tests can satisfy stack parameters without having to stub out all of the
discovery types and we are setup to change discovery implementation
details moving forward.

This requires updating the gateway and ingress stacks in addition to the
(default) sidecar stack.

Generally, we prefer duplicating similar target types to coupling
modules. In followup changes, we can introduce macros to reduce `Param`
boilerplate if necessary.

Changes include:

* A new `Gateway` builder type (analogous to `Inbound` and `Outbound`);
* A new `Gateway::server` stack handles discovery and protocol
  switching;
* Service cacheing is performed for the http and opaq stacks (to
  preserve load balancer state, etc) using only the information needed
  for those stacks;
* Concrete stacks provide the only data-plane queue. This enables
  outer stacks to be cloneable. This also means that the concrete stacks
  are the only stacks with stack metrics;
* The `switch_logical` module is removed;
* The `http::detect` module is moved into the `protocol` stack, which
  now supports discovery-hinted protocols to bypass detection.
* The `logical` and `endpoint` modules are removed in favor of
  protocol-specific modules.