integration: add destination update builders (#2232)
Currently, updates sent by the mock destination controller used in
integration tests are constructed by a number of free functions with
names like `destination_add_tls` and `destination_add_h2_hinted` and so
on. These functions determine different aspects of the generated
destination updates, such as protocol hints, labels, and TLS identity.
This works acceptably for the current integration tests. However, in
order to add new tests, it may be desirable to add additional
destination configurations. For example, for testing the handling of
various protocol hints on direct connections, it's necessary to set both
a protocol hint *and* a TLS identity (direct connections require TLS).
The current approach has separate functions for constructing updates
with a protocol hint and updates with TLS.

In order to make the construction of test destinations more flexible,
this PR replaces the current approach with a builder. This way, the
logic for constructing updates with different attributes can be composed
more easily, without having to implement separate constructors for the
entire matrix of (protocol hint, TLS identity, addr labels, addr set
labels) combinations.

This was factored out of PR #2209 so that the test framework changes can
be merged separately.