gateway: Fix URI form for gatewayed HTTP/1 requests
A recent change broke how gateways handle the URI of upgraded HTTP/1.1
requests. This problem was caused by normaling the URI twice (once in
the inbound HTTP server and once in the outbound HTTP server).

To fix this, the outbound HTTP server has been removed from the cached
stack so that it is not included in gatewayed services.

Furthermore, to test this, this change splits `push_http_client` from
the `Outbound::push_http_endpoint` stack and `push_http_tcp_server` from
the `Inbound::push_http_server` stack. This allows us to write tests
that mock HTTP-level services.