inbound: refactor newtypes for policy addresses
Depends on #2186.

Currently, the `linkerd-app-inbound` crate's `GetPolicy` trait performs
lookups for `OrigDstAddr`s. This is not strictly correct: the
`OrigDstAddr` newtype should specifically represent an address that was
returned by a `getsockopt(..., SO_ORIGINAL_DST, ...)` call, and in the
case of the inbound direct stack, the "original dst address" passed to
`GetPolicy` may actually be synthesized from a port override in the
`TransportHeader`. Therefore, it would be nicer if we had a `LookupAddr`
type specifically representing an address used as an argument for policy
lookups, analogous to the `profiles::LookupAddr` newtype used on the
outbound side. See [this comment][1] for details.

Therefore, this branch makes the following changes:

- Policy types (e.g. `ServerPolicy`, `AllowPermit`, etc) which store
  addresses represent those addresses as `ServerAddr` rather than
  `OrigDstAddr`, and perform address matching on `ServerAddr`s.
- The `inbound::policy` module now defines a `LookupAddr` newtype, which
  is passed as an argument to `GetPolicy`.
- The inbound accept and direct stacks provide `ExtractParam` impls to
  the `policy::Discover` layer indicating how `LookupAddr`s are produced
  from targets in that stack. This makes the different behaviors between
  the normal accept stack and the direct stack more explicit at the
  callsite.

[1]: https://github.com/linkerd/linkerd2-proxy/pull/2186#discussion_r1115153342